requirements.txt
neat==0.4.1
neat-python==0.92
numpy==1.16.2
opencv-python==4.1.0.25
matplotlib==3.0.3
pickleshare==0.7.5
gym-retro>=0.7.0
